backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
zeep>=4.2.1
httpx>=0.27.0
xmltodict>=0.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0
cryptography>=42.0.0
celery[redis]>=5.3.0
redis>=5.0.0
python-multipart>=0.0.9

# Dev / Test
pytest>=7.4.0
pytest-asyncio>=0.23.0
aiosqlite>=0.20.0
fakeredis>=2.21.0
respx>=0.21.0
